+++ src/reuse/project.py
@@ -231,12 +231,7 @@ class Project:
     def _licenses(self) -> Dict[str, Path]:
         """Return a dictionary of all licenses in the project, with their SPDX
         identifiers as names and paths as values.
-
-        If no name could be found for a license file, name it
-        "LicenseRef-Unknown0" and count upwards for every other unknown file.
         """
-
-        unknown_counter = 0
         license_files = dict()
 
         directory = str(self.root.resolve() / "LICENSES/**")
@@ -256,8 +251,7 @@ class Project:
             try:
                 identifier = self._identifier_of_license(path)
             except IdentifierNotFound:
-                identifier = "LicenseRef-Unknown{}".format(unknown_counter)
-                unknown_counter += 1
+                identifier = path.stem
                 _LOGGER.warning(
                     _(
                         "Could not resolve SPDX License Identifier of {path}, "