+++ CHANGELOG.md
@@ -29,6 +29,8 @@ The versions follow [semantic versioning](https://semver.org).
 
 - Added `--template` to `reuse addheader`.
 
+- `binaryornot` added as new dependency.
+
 ### Changed
 
 - `reuse addheader` now automatically adds the current year to the copyright
@@ -36,6 +38,8 @@ The versions follow [semantic versioning](https://semver.org).
 
 - `reuse addheader` preserves the original header below the new header if it did
   not contain any SPDX information.
+
+- `reuse addheader` now correctly handles `.license` files.
 
 ## 0.4.1 - 2019-08-07